% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-model.R
\name{all_composites}
\alias{all_composites}
\title{Get all composite constructs in a model}
\usage{
all_composites(seminr_model)
}
\arguments{
\item{seminr_model}{An estimated seminr model.}
}
\value{
A character vector of construct names.
}
\description{
Returns the names of constructs estimated as composites (i.e., not
reflective common factors) in an estimated seminr model.
}
