% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_mmMatrix.R
\name{all_construct_names}
\alias{all_construct_names}
\title{Get names of all constructs in a measurement model}
\usage{
all_construct_names(measurement_model)
}
\description{
Get names of all constructs in a measurement model
}
