% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_smMatrix.R
\name{all_endogenous}
\alias{all_endogenous}
\title{Get all endogenous construct names in a structural model}
\usage{
all_endogenous(smMatrix)
}
\description{
Get all endogenous construct names in a structural model
}
