% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_smMatrix.R
\name{all_exogenous}
\alias{all_exogenous}
\title{Get all exogenous construct names in a structural model}
\usage{
all_exogenous(smMatrix)
}
\description{
Get all exogenous construct names in a structural model
}
