% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_smMatrix.R
\name{antecedents_of}
\alias{antecedents_of}
\title{Get antecedent construct names for a give construct in a model}
\usage{
antecedents_of(outcome, smMatrix)
}
\description{
Get antecedent construct names for a give construct in a model
}
