% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_associations.R
\name{associations}
\alias{associations}
\title{SEMinR provides functions to specify inter-item covariances that
should be estimated, typically under CBSEM.}
\usage{
associations(...)
}
\description{
SEMinR provides functions to specify inter-item covariances that
should be estimated, typically under CBSEM.
}
\examples{
covaries <- associations(
  item_errors(c("a1", "a2"), c("b1", "b2")),
  item_errors("a3", "c3")
)
}
