% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_test_utils.R
\name{check_test_plot}
\alias{check_test_plot}
\title{A function to create regression plots (maybe not needed?)}
\usage{
check_test_plot(plot, title, plot_dir = "regression_plots", refresh = FALSE)
}
\arguments{
\item{plot}{the plot}

\item{title}{a unique title}

\item{plot_dir}{optional directory name}

\item{refresh}{whether to refresh all test cases}
}
\value{
TRUE if plots were the same, FALSE if they did not exist, or failed
}
\description{
A function to create regression plots (maybe not needed?)
}
