% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructs.R
\name{composite}
\alias{composite}
\title{Composite construct measurement model specification}
\usage{
measurement_model <- constructs(
                         composite("Satisfaction", multi_items("SAT", 1:3), weights = "correlation"),
                       )
}
\arguments{
\item{\code{construct_name}}{of construct}

\item{\code{item_names}}{returned by the \code{multi_items} or \code{single_item} functions}

\item{\code{weights}}{is the relationship between the construct and its items. This can be
specified as "correlation" or "A" for correlation weights(Mode A) or as "regression" or "B" for
regression weights(Mode B).}
}
\description{
\code{composite} creates the composite measurement model matrix for a specific construct,
specifying the relevant items of the construct and assigning the relationship of either
correlation weights (Mode A) or regression weights (Mode B).
}
\details{
This function conveniently maps composite defined measurement items to a construct and is
estimated using PLS.
}
\examples{
  mobi_mm <- constructs(
    composite("Image",        multi_items("IMAG", 1:5), weights = "correlation"),
    composite("Expectation",  multi_items("CUEX", 1:3), weights = "A"),
    composite("Quality",      multi_items("PERQ", 1:7), weights = "regression"),
    composite("Value",        multi_items("PERV", 1:2), weights = "B")
  )
}
\seealso{
See \code{\link{composite}}, \code{\link{reflective}}, \code{\link{multi_items}}
  and \code{\link{single_item}}
}

