% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_metrics.R
\name{compute_AVE}
\alias{compute_AVE}
\title{Returns Average Variance Extracted given construct's loadings}
\usage{
compute_AVE(lambdas)
}
\description{
Returns Average Variance Extracted given construct's loadings
}
