% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_construct_rhoA}
\alias{compute_construct_rhoA}
\title{Computes Henseler's rhoA}
\usage{
compute_construct_rhoA(weights, mmMatrix, construct, obsData)
}
\description{
Computes Henseler's rhoA
}
