% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_itcriteria_weights}
\alias{compute_itcriteria_weights}
\title{Function to calculate Akaike weights for IT Criteria}
\usage{
compute_itcriteria_weights(vector_of_itcriteria)
}
\arguments{
\item{vector_of_itcriteria}{This argument is a vector consisting of the IT criterion estimated
value for each model.}
}
\description{
Function to calculate Akaike weights for IT Criteria
}
