% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_metrics.R
\name{compute_rhoC}
\alias{compute_rhoC}
\title{Returns Composite Reliability given construct's loadings}
\usage{
compute_rhoC(lambdas)
}
\description{
Returns Composite Reliability given construct's loadings
}
