% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_vif}
\alias{compute_vif}
\title{Function to apply over manifests of a construct and return VIF values}
\usage{
compute_vif(target, predictors, model_data)
}
\description{
Function to apply over manifests of a construct and return VIF values
}
