% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-mmMatrix.R
\name{construct_name}
\alias{construct_name}
\title{Get the name of a single construct specification}
\usage{
construct_name(construct)
}
\arguments{
\item{construct}{A `construct` vector.}
}
\value{
A character string with the construct name.
}
\description{
Returns the construct's name from a `construct` vector (the user-side
specification produced by `reflective()`, `composite()`, etc.).
}
