% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_metrics.R
\name{cor_vifs}
\alias{cor_vifs}
\title{lm_vif <- function(data, model) {
  1/(1-summary(lm(model, as.data.frame(data)))$r.squared)
}}
\usage{
cor_vifs(cor_matrix, iv_names)
}
\description{
> lm_vif(scores, Value ~ Satisfaction + Complaints)
[1] 1.769965
> lm_vif(scores, Satisfaction ~ Value + Complaints)
[1] 2.455003
> lm_vif(scores, Complaints ~ Value + Satisfaction)
[1] 1.6049
}
\details{
> cor_vifs(cors, c("Value", "Satisfaction", "Complaints"))
   Value Satisfaction   Complaints
1.769965     2.455003     1.604900
}
