% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{create_theme}
\alias{create_theme}
\title{Factory for creating a theme object}
\usage{
create_theme(
  item_style = createItemStyle(),
  construct_style = createConstructStyle(),
  outer_weight_style = createOuterWeightStyle(),
  inner_weight_style = createInnerWeightStyle()
)
}
\arguments{
\item{item_style}{the dot-style for an item}

\item{construct_style}{the dot-style for a construct}

\item{outer_weight_style}{the dot-style for the measurement model}

\item{inner_weight_style}{the dot-style for the structural model}
}
\value{
a theme object used in the plot_model function
}
\description{
Factory for creating a theme object
}
