% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{createConstructStyle}
\alias{createConstructStyle}
\title{Creates a style definition for measurement constructs}
\usage{
createConstructStyle(
  color = "black",
  fill = "white",
  fontsize = 12,
  height = 0.5,
  width = 1
)
}
\arguments{
\item{color}{Color of node}

\item{fill}{fill of the node}

\item{fontsize}{Size of the text}

\item{height}{the height of the node}

\item{width}{the width of the node}
}
\value{
dot-String for styling nodes
}
\description{
Creates a style definition for measurement constructs
}
