% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{createEdgeStyle}
\alias{createEdgeStyle}
\title{Creates a style definition for edges}
\usage{
createEdgeStyle(
  color = "black",
  fontsize = 7,
  fontname = "helvetica",
  forward = TRUE,
  minlen = NA
)
}
\arguments{
\item{color}{Color of edge}

\item{fontsize}{Size of the text}

\item{fontname}{Font-type}

\item{forward}{direction of arrow}

\item{minlen}{minimal length of line}
}
\value{
dot-String for styling edges
}
\description{
Creates a style definition for edges
}
