% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{createInnerWeightStyle}
\alias{createInnerWeightStyle}
\title{Function to determine the edges for a structural model}
\usage{
createInnerWeightStyle(color = "black", fontsize = 9, forward = TRUE)
}
\arguments{
\item{color}{color of edge}

\item{fontsize}{Size of text}

\item{forward}{direction of arrow}

\item{minlen}{minimal length of arrow}
}
\value{
dot-string for styling edges
}
\description{
Function to determine the edges for a structural model
}
