% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{createItemStyle}
\alias{createItemStyle}
\title{Creates a style definition for measurement items}
\usage{
createItemStyle(
  color = "dimgray",
  fill = "white",
  fontsize = 8,
  height = 0.2,
  width = 0.4
)
}
\arguments{
\item{color}{Color of node}

\item{fill}{fill of the node}

\item{fontsize}{Size of the text}

\item{height}{the height of the node}

\item{width}{the width of the node}
}
\value{
dot-String for styling nodes
}
\description{
Creates a style definition for measurement items
}
