% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{createOuterWeightStyle}
\alias{createOuterWeightStyle}
\title{Function to determine the edges for a measurement model}
\usage{
createOuterWeightStyle(
  color = "dimgray",
  fontsize = 7,
  forward = FALSE,
  minlen = 1
)
}
\arguments{
\item{color}{color of edge}

\item{fontsize}{Size of text}

\item{forward}{direction of arrow}

\item{minlen}{minimal length of arrow}
}
\value{
dot-string for styling edges
}
\description{
Function to determine the edges for a measurement model
}
