% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{createStyle}
\alias{createStyle}
\title{Creates a style definition for nodes}
\usage{
createStyle(
  shape = "box",
  color = "dimgray",
  fill = "white",
  fontsize = 8,
  height = 0.2,
  width = 0.4,
  fontname = "helvetica"
)
}
\arguments{
\item{shape}{the shape for the node}

\item{color}{Color of node}

\item{fill}{fill of the node}

\item{fontsize}{Size of the text}

\item{height}{the height of the node}

\item{width}{the width of the node}

\item{fontname}{Font-type}
}
\value{
dot-String for styling nodes
}
\description{
Creates a style definition for nodes
}
