% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_lavaan.R
\name{curated_fit_metrics}
\alias{curated_fit_metrics}
\title{Returns selected fit metrics: ordinary, robustable, robust, and scaled}
\usage{
curated_fit_metrics(fit_metrics)
}
\description{
Returns selected fit metrics: ordinary, robustable, robust, and scaled
}
