% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addin.R
\name{custom_theme_addin}
\alias{custom_theme_addin}
\title{Starts the model viewer in RStudio}
\usage{
custom_theme_addin()
}
\value{
the updated dot code
}
\description{
Starts the model viewer in RStudio
}
