% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{default_edge_template}
\alias{default_edge_template}
\title{The default template for labeling bootstrapped edges}
\usage{
default_edge_template()
}
\value{
The default string
}
\description{
The default template for labeling bootstrapped edges
}
