% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_data_structures.R
\name{df_xtab_matrix}
\alias{df_xtab_matrix}
\title{Cross-tabulates columns of a dataframe into a matrix with NAs for unspecified pairs}
\usage{
df_xtab_matrix(model, df, rows, columns)
}
\arguments{
\item{model}{A \code{formula} indicating relevant columns from data frame}

\item{df}{A \code{data.frame} of columns to cross-tabulate}

\item{rows}{A \code{vector} of row names for the matrix to sort by}

\item{columns}{A \code{vector} of column names for the matrix to sort by}
}
\description{
Cross-tabulates columns of a dataframe into a matrix with NAs for unspecified pairs
}
