% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{dot_component_mm}
\alias{dot_component_mm}
\title{Generates the dot code for the measurement model}
\usage{
dot_component_mm(model, theme)
}
\arguments{
\item{model}{the model to use}

\item{theme}{the theme to use}
}
\description{
Generates the dot code for the measurement model
}
