% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_dot.R
\name{dot_graph.cbsem_model}
\alias{dot_graph.cbsem_model}
\title{Plotting of covariance based SEMs models using semPLOT}
\usage{
\method{dot_graph}{cbsem_model}(
  model,
  title = "",
  theme = NULL,
  what = "std",
  whatLabels = "std",
  ...
)
}
\arguments{
\item{model}{the CBSEM model}

\item{title}{Unused}

\item{theme}{Unused}

\item{what}{The metric to use for edges ("path", "est", "std", "eq", "col")}

\item{whatLabels}{The metric to use for edge labels}

\item{...}{Parameters passed to the \link[semPlot]{semPaths} function}
}
\description{
For a full description of parameters see \link[semPlot]{semPaths}
}
