% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_dot.R
\name{dot_graph.measurement_model}
\alias{dot_graph.measurement_model}
\title{Convert a seminr measurement model to a Graphviz representation}
\usage{
dot_graph.measurement_model(measurement_model, title, theme = NULL)
}
\arguments{
\item{measurement_model}{Model created with \code{seminr}.}

\item{title}{An optional title for the plot}

\item{theme}{Theme created with \code{\link{create_theme}}.}
}
\value{
The path model as a formatted string in dot language.
}
\description{
The function also tries to run a CFA on the measurement model and uses the loadings
in plots.
With the help of the \code{DiagrammeR} package this code can then be plotted in
various contexts.
}
\details{
Current limitations:
- experimental
}
\examples{
mobi <- mobi

#seminr syntax for creating measurement model
mobi_mm <- constructs(
             reflective("Image",        multi_items("IMAG", 1:5)),
             reflective("Expectation",  multi_items("CUEX", 1:3)),
             reflective("Quality",      multi_items("PERQ", 1:7)),
             reflective("Value",        multi_items("PERV", 1:2)),
             reflective("Satisfaction", multi_items("CUSA", 1:3)),
             reflective("Complaints",   single_item("CUSCO")),
             reflective("Loyalty",      multi_items("CUSL", 1:3))
           )
res <- dot_graph.measurement_model(mobi_mm, title = "Example plot")

\dontrun{
DiagrammeR::grViz(res)
}
}
