% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{dot_graph.structural_model}
\alias{dot_graph.structural_model}
\title{Convert a seminr measurement model to a Graphviz representation}
\usage{
\method{dot_graph}{structural_model}(model, title = "", theme = NULL, ...)
}
\arguments{
\item{model}{Model created with \code{seminr}.}

\item{title}{An optional title for the plot}

\item{theme}{Theme created with \code{\link{seminr_theme_create}}.}

\item{...}{Unused}
}
\description{
Convert a seminr measurement model to a Graphviz representation
}
\examples{
# - - - - - - - - - - - - - - - -
# Example for plotting a structural model
mobi_sm <- relationships(
  paths(from = "Image",        to = c("Expectation", "Satisfaction", "Loyalty")),
  paths(from = "Expectation",  to = c("Quality", "Value", "Satisfaction")),
  paths(from = "Quality",      to = c("Value", "Satisfaction")),
  paths(from = "Value",        to = c("Satisfaction")),
  paths(from = "Satisfaction", to = c("Complaints", "Loyalty")),
  paths(from = "Complaints",   to = "Loyalty")
)
res <- dot_graph(mobi_sm, title = "Preview structural model")

\dontrun{
DiagrammeR::grViz(res)
}
}
