% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{dot_subcomponent_mm}
\alias{dot_subcomponent_mm}
\title{generates the dot code for a subgraph (per construct)}
\usage{
dot_subcomponent_mm(index, model, theme)
}
\arguments{
\item{index}{the index of the construct}

\item{model}{the model to use}

\item{theme}{the theme to use}
}
\description{
generates the dot code for a subgraph (per construct)
}
