% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{edge_template_default}
\alias{edge_template_default}
\title{The default template for labeling bootstrapped edges}
\usage{
edge_template_default()
}
\value{
The template string
}
\description{
The default template for labeling bootstrapped edges
}
