% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{edge_template_minimal}
\alias{edge_template_minimal}
\title{A minimal template for labeling bootstrapped edges that only shows the
bootstrapped mean value}
\usage{
edge_template_minimal()
}
\value{
The template string
}
\description{
A minimal template for labeling bootstrapped edges that only shows the
bootstrapped mean value
}
