% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{seminr Model Estimation Function}
\usage{
estimate(seminr_model, ...)
}
\arguments{
\item{seminr_model}{A \code{plsm} object generated by the \code{seminr::model} function}

\item{...}{A list of parameters inherited from the estimation function (default is semPLS)}
}
\description{
The \code{seminr} package provides a natural syntax for researchers to describe PLS
structural equation models. \code{seminr} is compatible with semPLS, simplePLS and matrixPLS,
meaning that once a model is specified, it can be used across models for comparison.
\code{seminr::estimate} provides the verb for estimating a pls model from the plsm object
generated by \code{seminr::model}.
}
\examples{
data("mobi", package = "semPLS")

#seminr syntax for creating measurement model
mobi_mm <- measure(
             reflect("Image",        multi_items("IMAG", 1:5)),
             reflect("Expectation",  multi_items("CUEX", 1:3)),
             reflect("Quality",      multi_items("PERQ", 1:7)),
             reflect("Value",        multi_items("PERV", 1:2)),
             reflect("Satisfaction", multi_items("CUSA", 1:3)),
             reflect("Complaints",   single_item("CUSCO")),
             reflect("Loyalty",      multi_items("CUSL", 1:3))
           )
#seminr syntax for creating structural model
mobi_sm <- structure(
  paths(from = "Image",        to = c("Expectation", "Satisfaction", "Loyalty")),
  paths(from = "Expectation",  to = c("Quality", "Value", "Satisfaction")),
  paths(from = "Quality",      to = c("Value", "Satisfaction")),
  paths(from = "Value",        to = c("Satisfaction")),
  paths(from = "Satisfaction", to = c("Complaints", "Loyalty")),
  paths(from = "Complaints",   to = "Loyalty")
)

plsm_model <- model(data = mobi,
                    measurement_model = mobi_mm,
                    structural_model = mobi_sm)

mobi_pls <- estimate(plsm_model)

print_paths(mobi_pls)
plot_scores(mobi_pls)
}
\seealso{
\code{\link{structure}} \code{\link{measure}} \code{\link{paths}} \code{\link{interact}}
         \code{\link{estimate}}
}

