% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_factor_scores.R
\name{estimate_lavaan_ten_berge}
\alias{estimate_lavaan_ten_berge}
\title{seminr estimate_lavaan_ten_berge() function}
\usage{
estimate_lavaan_ten_berge(fit)
}
\arguments{
\item{fit}{A fitted \code{lavaan} object – can be extracted from cbesem estimation or from using Lavaan directly.}
}
\value{
A list with two elements: ten berge scores; weights for calculating scores
}
\description{
Estimates factor scores using ten Berge method for a fitted Lavaan model
}
\examples{
cbsem <- estimate_cbsem(...)
tb <- estimate_lavaan_ten_berge(cbsem$lavaan_model)
tb$scores
tb$weights

}
