% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pls_mga.R
\name{estimate_pls_mga}
\alias{estimate_pls_mga}
\title{Performs PLS-MGA to report significance of path differences between two subgroups of data}
\usage{
estimate_pls_mga(pls_model, condition, nboot = 2000)
}
\arguments{
\item{pls_model}{SEMinR PLS model estimated on the full sample}

\item{condition}{logical vector of TRUE/FALSE indicating which rows of sample data are in group 1}

\item{nboot}{number of bootstrap resamples to use in PLS-MGA

TODO: parallelize param}
}
\description{
Performs PLS-MGA to report significance of path differences between two subgroups of data
}
