% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{extract_bootstrapped_values}
\alias{extract_bootstrapped_values}
\title{extract bootstrapped statistics from an edge using a row_index}
\usage{
extract_bootstrapped_values(ltbl, row_index, model, theme)
}
\arguments{
\item{ltbl}{a table of bootstrapped values (weights, loadings, path coefficients)}

\item{row_index}{the index for the specific edge to extract}

\item{model}{the model to use}

\item{theme}{the theme to use}
}
\description{
extract bootstrapped statistics from an edge using a row_index
}
