% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_htmt.R
\name{extract_htmt_nodes}
\alias{extract_htmt_nodes}
\title{Helper function that applies formatting to each construct}
\usage{
extract_htmt_nodes(model, theme)
}
\arguments{
\item{model}{the model to use}

\item{theme}{the theme to use}
}
\value{
Returns a string of the structural model in dot notation.
}
\description{
Helper function that applies formatting to each construct
}
