% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{extract_mm_coding}
\alias{extract_mm_coding}
\title{extracts the constructs and their types from the model}
\usage{
extract_mm_coding(model)
}
\arguments{
\item{model}{the model to use}
}
\description{
extracts the constructs and their types from the model
}
