% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{extract_mm_edge_value}
\alias{extract_mm_edge_value}
\title{gets the mm_edge value (loading, weight) for bootstrapped and regular models}
\usage{
extract_mm_edge_value(model, theme, indicator, construct)
}
\arguments{
\item{model}{the model to use}

\item{theme}{the theme to use}

\item{indicator}{the indicator to use}

\item{construct}{the construct to use}
}
\description{
gets the mm_edge value (loading, weight) for bootstrapped and regular models
}
