% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{extract_mm_edges}
\alias{extract_mm_edges}
\title{extract mm edges from model for a given index of all constructs}
\usage{
extract_mm_edges(index, model, theme, weights = 1000)
}
\arguments{
\item{index}{the index of the construct}

\item{model}{the model to use}

\item{theme}{the theme to use}

\item{weights}{a default weight for measurment models (high values suggested)}
}
\description{
extract mm edges from model for a given index of all constructs
}
