% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{extract_mm_nodes}
\alias{extract_mm_nodes}
\title{gets the individual nodes and applies formatting}
\usage{
extract_mm_nodes(index, model, theme)
}
\arguments{
\item{index}{the index of the construct}

\item{model}{the model to use}

\item{theme}{the theme to use}
}
\description{
gets the individual nodes and applies formatting
}
