% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{fl_criteria_table}
\alias{fl_criteria_table}
\title{fl_criteria_table can be used to generate simple and effective table for checking Fornel Larcker criteria.}
\usage{
fl_criteria_table(sum_mobi_pls)
}
\arguments{
\item{summary_object}{A \code{summary.seminr_model} containing the summarized seminr model.}
}
\description{
The \code{fl_criteria_table} function generates a table of correaltions and AVE values for inspection.
}
\examples{
#seminr syntax for creating measurement model
mobi_mm <- constructs(
             reflective("Image",        multi_items("IMAG", 1:5)),
             reflective("Expectation",  multi_items("CUEX", 1:3)),
             reflective("Quality",      multi_items("PERQ", 1:7)),
             reflective("Value",        multi_items("PERV", 1:2)),
             reflective("Satisfaction", multi_items("CUSA", 1:3)),
             reflective("Complaints",   single_item("CUSCO")),
             reflective("Loyalty",      multi_items("CUSL", 1:3))
           )
#seminr syntax for creating structural model
mobi_sm <- relationships(
  paths(from = "Image",        to = c("Expectation", "Satisfaction", "Loyalty")),
  paths(from = "Expectation",  to = c("Quality", "Value", "Satisfaction")),
  paths(from = "Quality",      to = c("Value", "Satisfaction")),
  paths(from = "Value",        to = c("Satisfaction")),
  paths(from = "Satisfaction", to = c("Complaints", "Loyalty")),
  paths(from = "Complaints",   to = "Loyalty")
)

mobi_pls <- estimate_pls(data = mobi,
                           measurement_model = mobi_mm,
                           structural_model = mobi_sm)

sum_mobi_pls <- summary(mobi_pls)
fl_criteria_table(sum_mobi_pls)

}
\references{
Hair, J. F., Hult, G. T. M., Ringle, C. M., and Sarstedt, M. (2017). A Primer on Partial Least Squares
 Structural Equation Modeling (PLS-SEM), 2nd Ed., Sage: Thousand Oaks.
}
