% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{format_endo_node_label}
\alias{format_endo_node_label}
\title{Helps to render a node label for endogenous variables}
\usage{
format_endo_node_label(theme, name, rstring)
}
\arguments{
\item{theme}{the theme to use}

\item{name}{the content of the name placeholder}

\item{rstring}{the content of the rstring placeholder}
}
\value{
Returns the formatted string
}
\description{
Helps to render a node label for endogenous variables
}
