% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{format_exo_node_label}
\alias{format_exo_node_label}
\title{Helps to render a node label for exogenous variables}
\usage{
format_exo_node_label(theme, name)
}
\arguments{
\item{theme}{the theme to use}

\item{name}{the content of the name placeholder}
}
\value{
Returns the formatted string
}
\description{
Helps to render a node label for exogenous variables
}
