% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{get_construct_type}
\alias{get_construct_type}
\title{Returns the type of a construct from a model}
\usage{
get_construct_type(model, construct)
}
\arguments{
\item{model}{the model to get the type from}

\item{construct}{the character string name of the construct}
}
\value{
Returns a character string
}
\description{
Returns the type of a construct from a model
}
