% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{get_manifest_element_size}
\alias{get_manifest_element_size}
\title{Gets the optimal size for manifest elements in the plot}
\usage{
get_manifest_element_size(model, theme)
}
\arguments{
\item{model}{the model to use}

\item{theme}{the theme to use}
}
\value{
Returns a two-element vector with c(width, height)
}
\description{
Currently orients on reflective theme settings
}
