% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{get_mm_edge_style}
\alias{get_mm_edge_style}
\title{individual styles for measurement model edges}
\usage{
get_mm_edge_style(theme, construct_type, flip = FALSE)
}
\arguments{
\item{theme}{the theme to use}

\item{construct_type}{Forward direction?}

\item{flip}{invert the arrow direction because of sink?}
}
\description{
individual styles for measurement model edges
}
