% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{get_mm_node_style}
\alias{get_mm_node_style}
\title{get global measurement model node style}
\usage{
get_mm_node_style(theme)
}
\arguments{
\item{theme}{the theme to use}
}
\description{
get global measurement model node style
}
