% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{get_sm_node_shape}
\alias{get_sm_node_shape}
\title{Get a string to insert into a node specification using the themed shape}
\usage{
get_sm_node_shape(model, construct, theme)
}
\arguments{
\item{model}{the model to use}

\item{construct}{the construct to use}

\item{theme}{the theme to use}
}
\value{
Returns a string that determines the shape of a node
}
\description{
Get a string to insert into a node specification using the themed shape
}
