% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_utils.R
\name{get_theme_doc}
\alias{get_theme_doc}
\title{Get Documentation for theme options to use in Addin}
\usage{
get_theme_doc()
}
\value{
Returns a data.frame with the documented arguments
}
\description{
Get Documentation for theme options to use in Addin
}
\keyword{internal}
