% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{get_value_dependent_mm_edge_style}
\alias{get_value_dependent_mm_edge_style}
\title{Formats the style of the structural model edges}
\usage{
get_value_dependent_mm_edge_style(value, theme)
}
\arguments{
\item{value}{value to compare for negativity}

\item{theme}{the theme to use}
}
\value{
Returns the style for the edge (both style and color)
}
\description{
Formats the style of the structural model edges
}
