% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seminr_plot.R
\name{getFormattedNode}
\alias{getFormattedNode}
\title{Generates the dot_code for a node}
\usage{
getFormattedNode(construct, model, theme)
}
\arguments{
\item{construct}{Name of the construct to generate the dot code for}

\item{model}{the model object that contains the r^2 values}

\item{theme}{the theme that is being used for plotting}
}
\value{
Dot-code for the construct
}
\description{
Generates the dot_code for a node
}
