% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interact}
\alias{interact}
\alias{interaction_ortho,}
\alias{interaction_scaled}
\title{Interaction Functions}
\usage{
# Create two new interactions (factor1.factor2 and factor1.factor3) using

 # (1) standardized product indicator approach as per Henseler & Chin (2010):
 interact( interaction_scaled("factor1", "factor2"),
           interaction_scaled("factor1", "factor3"))

 # (2) orthogonalization approach as per Henseler & CHin (2010):
 interact( interaction_ortho("factor1", "factor2"),
           interaction_ortho("factor1", "factor3"))
}
\arguments{
\item{...}{Interaction Combinations as generated by the \code{interaction_scaled} or
\code{interaction_ortho} methods.}
}
\description{
\code{interact} creates interaction measurement items by multipying all combination of factor items.
}
\details{
This function automatically generates interaction measurement items for a PLS SEM.
}
\examples{
data("mobi", package = "semPLS")

mobi_xm <- interact(
  interaction_ortho("Image", "Expectation"),
  interaction_ortho("Image", "Value")
)

mobi_pls <- estimate_pls(mobi, mobi_mm, mobi_xm, mobi_sm)
print_paths(mobi_pls)

}

