% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interaction_ortho}
\alias{interaction_ortho}
\title{\code{interaction_ortho} creates interaction measurement items by using the orthogonalized approach..}
\usage{
# orthogonalization approach as per Henseler & CHin (2010):
 interaction_ortho(factor1, factor2)
}
\arguments{
\item{factor1}{The first factor which is subject to the interaction.}

\item{factor2}{The second factor which is subject to the interaction.}
}
\description{
This function automatically generates interaction measurement items for a PLS SEM using the orthogonalized approach..
}
\examples{
data(mobi)

# seminr syntax for creating measurement model
mobi_mm <- constructs(
  composite("Image",        multi_items("IMAG", 1:5)),
  composite("Expectation",  multi_items("CUEX", 1:3)),
  composite("Value",        multi_items("PERV", 1:2)),
  composite("Satisfaction", multi_items("CUSA", 1:3))
)
mobi_xm <- interactions(
  interaction_ortho("Image", "Expectation"),
  interaction_ortho("Image", "Value")
)

#  structural model: note that name of the interactions factor should be
#  the names of its two main factors joined by a '.' in between.
mobi_sm <- relationships(
  paths(to = "Satisfaction",
        from = c("Image", "Expectation", "Value",
                 "Image.Expectation", "Image.Value"))
)

mobi_pls <- estimate_pls(mobi, mobi_mm, mobi_xm, mobi_sm)
print_paths(mobi_pls)

}
\references{
Henseler, J., & Chin, W. W. (2010), A comparison of approaches for the analysis of interaction effects
 between latent variables using partial least squares path modeling. Structural Equation Modeling, 17(1), 82–109.
 https://doi.org/10.1080/10705510903439003
}
