% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interaction_scaled}
\alias{interaction_ortho,}
\alias{interaction_scaled}
\title{\code{interaction_scaled} creates interaction measurement items by multipying all combination of factor items.}
\usage{
# Create two new interactions (factor1.factor2 and factor1.factor3) using

 # standardized product indicator approach as per Henseler & Chin (2010):
 interactions( interaction_scaled("factor1", "factor2"),
           interaction_scaled("factor1", "factor3"))
}
\arguments{
\item{...}{Interaction Combinations as generated by the \code{interaction_scaled} method.}
}
\description{
This function automatically generates interaction measurement items for a PLS SEM.
}
\examples{
data("mobi", package = "semPLS")

mobi_xm <- interactions(
  interaction_scaled("Image", "Expectation"),
  interaction_scaled("Image", "Value")
)

mobi_pls <- estimate_pls(mobi, mobi_mm, mobi_xm, mobi_sm)
print_paths(mobi_pls)

}

