% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{is.seminr_theme}
\alias{is.seminr_theme}
\title{Reports whether x is a seminr_theme object}
\usage{
is.seminr_theme(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a seminr_theme object
}
\keyword{internal}
