% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{is_sink}
\alias{is_sink}
\title{Tests whether the i_th construct is endogenous or not}
\usage{
is_sink(model, index)
}
\arguments{
\item{model}{the model object}

\item{index}{the index of the construct to test}
}
\value{
whether the construct is endogenous or not
}
\description{
Tests whether the i_th construct is endogenous or not
}
