% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_associations.R
\name{item_errors}
\alias{item_errors}
\title{Specifies pair of items, or sets of items, that should covary. Used to
specify error covariances for \code{\link{associations}}.}
\usage{
item_errors(items_a, items_b)
}
\arguments{
\item{items_a}{One or more items that should covary}

\item{items_b}{One or more items that should covary}
}
\value{
A vector of items that covary.
}
\description{
Specifies pair of items, or sets of items, that should covary. Used to
specify error covariances for \code{\link{associations}}.
}
\examples{
item_errors(c("a1", "a2"), c("b1", "b2"))
}
