% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePLS.R
\name{items_per_mode}
\alias{items_per_mode}
\title{Create a function to get all the items of a given measurement mode for a given latent}
\usage{
items_per_mode(latent, mode, mmMatrix)
}
\description{
Create a function to get all the items of a given measurement mode for a given latent
}

