% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lastplot.R
\name{last_seminr_plot}
\alias{last_seminr_plot}
\title{Retrieve the last plot to be modified or created.}
\usage{
last_seminr_plot()
}
\description{
Retrieve the last plot to be modified or created.
}
\seealso{
[save_plot()]
}
\keyword{internal}
