% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_syntax.R
\name{lavaan_association}
\alias{lavaan_association}
\title{Create Lavaan syntax for a single association between items}
\usage{
lavaan_association(pair)
}
\description{
Create Lavaan syntax for a single association between items
}
