% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_syntax.R
\name{lavaan_construct}
\alias{lavaan_construct}
\title{Create lavaan syntax for a single construct's measurement}
\usage{
lavaan_construct(construct_matrix)
}
\description{
Create lavaan syntax for a single construct's measurement
}
