% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_syntax.R
\name{lavaan_item_associations}
\alias{lavaan_item_associations}
\title{Create Lavaan syntax for inter-item associations}
\usage{
lavaan_item_associations(item_associations)
}
\description{
Create Lavaan syntax for inter-item associations
}
