% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_syntax.R
\name{lavaan_mm_syntax}
\alias{lavaan_mm_syntax}
\title{Create Lavaan syntax for entire measurement model}
\usage{
lavaan_mm_syntax(mmMatrix)
}
\description{
Create Lavaan syntax for entire measurement model
}
