% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_syntax.R
\name{lavaan_regression}
\alias{lavaan_regression}
\title{Create lavaan syntax for a single construct's endogenous paths}
\usage{
lavaan_regression(outcome, smMatrix)
}
\description{
Create lavaan syntax for a single construct's endogenous paths
}
