% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_syntax.R
\name{lavaan_sm_syntax}
\alias{lavaan_sm_syntax}
\title{Create Lavaan syntax for entire structural model}
\usage{
lavaan_sm_syntax(structural_model)
}
\description{
Create Lavaan syntax for entire structural model
}
