% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_syntax.R
\name{lavaanify_name}
\alias{lavaanify_name}
\title{Makes sure construct name is valid for lavaan, or else transforms it}
\usage{
lavaanify_name(name)
}
\description{
Makes sure construct name is valid for lavaan, or else transforms it
}
