% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement.R
\name{measure}
\alias{measure}
\title{Measurement functions}
\usage{
measurement_model <- measure(
                         reflective("Satisfaction", multi_items("SAT", 1:3)),
                         composite("Reputation", multi_items("REP", 1:3)),
                         causal("Performance", c("ROE", "ROA"))
                       )
}
\arguments{
\item{\code{vector}}{of latent variable measurement specifications, as generated by the \code{reflective()}, \code{causal()}, or \code{composite()}}
}
\description{
\code{measure} creates the factors from measurement items by assigning the
relevant items to each factor and specifying reflective or formative (composite/causal) measurement models
}
\details{
This function conveniently maps measurement items to factors using
root name, numbers, and affixes with explicit definition of formative
or reflective structure
}
\examples{
  mobi_mm <- measure(
    reflective("Image",        multi_items("IMAG", 1:5)),
    reflective("Expectation",  multi_items("CUEX", 1:3)),
    reflective("Quality",      multi_items("PERQ", 1:7)),
    reflective("Value",        multi_items("PERV", 1:2)),
    reflective("Satisfaction", multi_items("CUSA", 1:3)),
    reflective("Complaints",   single_item("CUSCO")),
    reflective("Loyalty",      multi_items("CUSL", 1:3))
  )
}
\seealso{
See \code{\link{causal}}, \code{\link{composite}}, \code{\link{reflective}}, \code{\link{multi_items}}
  and \code{\link{single_item}}
}

