% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_mmMatrix.R
\name{mm2matrix}
\alias{mm2matrix}
\title{Convert measurement model into mmMatrix
- if measurement model is a matrix, return it directly (used in 2-stage)}
\usage{
mm2matrix(measurement_model)
}
\description{
Convert measurement model into mmMatrix
- if measurement model is a matrix, return it directly (used in 2-stage)
}
