% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_mmMatrix.R
\name{mm_interactions}
\alias{mm_interactions}
\title{Extract only interaction closures from measurement model}
\usage{
mm_interactions(measurement_model)
}
\description{
Extract only interaction closures from measurement model
}
