% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{mode_A}
\alias{mode_A}
\alias{mode_A,}
\alias{correlation_weights}
\title{Outer weighting scheme functions to estimate construct weighting.}
\usage{
mode_A(mmMatrix, i, normData, fscores)
}
\arguments{
\item{mmMatrix}{is the \code{measurement_model} - a source-to-target matrix representing the measurement model,
generated by \code{constructs}.}

\item{i}{is the name of the construct to be estimated.}

\item{normData}{is the dataframe of the normalized item data.}

\item{fscores}{is the matrix of construct scores generated by \code{simplePLS}.}
}
\description{
\code{mode_A}, \code{correlation_weights} and \code{mode_B}, \code{regression_weights} specify the outer weighting
scheme to be used in the estimation of the construct weights and score.
}
