% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{mode_plsc}
\alias{mode_plsc}
\title{Outer weighting scheme functions to estimate construct weighting.}
\usage{
mode_plsc(mmMatrix, j, normData, construct_scores)
}
\arguments{
\item{mmMatrix}{is the \code{measurement_model} - a source-to-target matrix
representing the measurement model, generated by \code{constructs}.}

\item{j}{is the name of the construct to be estimated.}

\item{normData}{is the dataframe of the normalized item data.}

\item{construct_scores}{is the matrix of construct scores generated by \code{estimate_model}.}
}
\value{
A matrix of estimated measurement model relations.
}
\description{
\code{mode_plsc} specifies outer weighting as mode A and then applies PLSc disattenuation
to mimic common-factor model estimates.
}
