% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{node_template_default}
\alias{node_template_default}
\title{The default template for labeling construct nodes}
\usage{
node_template_default()
}
\value{
The template string
}
\description{
The default template for labeling construct nodes
}
