% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_mmMatrix.R
\name{number_of_items}
\alias{number_of_items}
\title{get number of items from a construct in a measurement model}
\usage{
number_of_items(construct)
}
\description{
get number of items from a construct in a measurement model
}
