% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_utils.R
\name{package_docs}
\alias{package_docs}
\title{Create a list of section content from every documentation page from a package}
\usage{
package_docs(pkg)
}
\arguments{
\item{pkg}{String naming the package}
}
\value{
A list whose names are documentation page titles. Each element is
  itself a list of the details and sections for a page.
}
\description{
Create a list of section content from every documentation page from a package
}
\keyword{internal}
