% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_dot.R
\name{plot_model}
\alias{plot_model}
\title{Plot various SEMinR models}
\usage{
plot_model(model, title = "", theme = NULL, ...)
}
\arguments{
\item{model}{The model description}

\item{title}{An optional title for the plot}

\item{theme}{Theme created with \code{\link{seminr_theme_create}}.}

\item{...}{Additional parameters}
}
\value{
The path model as a formatted string in dot language.
}
\description{
With the help of the \code{DiagrammeR} package this dot graph can then be plotted in
various in RMarkdown, shiny, and other contexts.
Depending on the type of model, different parameters can be used.
Please check the \code{\link{dot_graph}} function for additional parameters
}
