% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_reliability}
\alias{plot_reliability}
\title{Plot Reliability can be used to generate simple and effective bar plots for reliability stats.}
\usage{
plot_reliability(sum_mobi_pls, "rhoA", 0.7))
}
\arguments{
\item{summary_object}{A \code{summary.seminr_model} containing the summarized seminr model.}

\item{criteria}{The reliability criteria to be viasualizes, AVE, RhoA, RhoC, or alpha.}

\item{limit}{The upper limit for reliability such as 0.7 for RhoA and 0.5 for AVE.}
}
\description{
The \code{plot_reliability} function generates a bar plot of reliability measures with a specified cut-off.
}
\examples{
#seminr syntax for creating measurement model
mobi_mm <- constructs(
             reflective("Image",        multi_items("IMAG", 1:5)),
             reflective("Expectation",  multi_items("CUEX", 1:3)),
             reflective("Quality",      multi_items("PERQ", 1:7)),
             reflective("Value",        multi_items("PERV", 1:2)),
             reflective("Satisfaction", multi_items("CUSA", 1:3)),
             reflective("Complaints",   single_item("CUSCO")),
             reflective("Loyalty",      multi_items("CUSL", 1:3))
           )
#seminr syntax for creating structural model
mobi_sm <- relationships(
  paths(from = "Image",        to = c("Expectation", "Satisfaction", "Loyalty")),
  paths(from = "Expectation",  to = c("Quality", "Value", "Satisfaction")),
  paths(from = "Quality",      to = c("Value", "Satisfaction")),
  paths(from = "Value",        to = c("Satisfaction")),
  paths(from = "Satisfaction", to = c("Complaints", "Loyalty")),
  paths(from = "Complaints",   to = "Loyalty")
)

mobi_pls <- estimate_pls(data = mobi,
                           measurement_model = mobi_mm,
                           structural_model = mobi_sm)

sum_mobi_pls <- summary(mobi_pls)
plot_reliability(sum_mobi_pls, "rhoA", 0.7)

}
\references{
Hair, J. F., Hult, G. T. M., Ringle, C. M., and Sarstedt, M. (2017). A Primer on Partial Least Squares
 Structural Equation Modeling (PLS-SEM), 2nd Ed., Sage: Thousand Oaks.
}
