% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistent_interaction.R
\name{PLSc_interact}
\alias{PLSc_interact}
\title{SEMinR - Consistent Interaction Function (experimental)}
\usage{
PLSc_interact(seminr_model)
}
\arguments{
\item{plsModel}{A \code{seminr_model} containing the estimated seminr model with only the two
antecedent constructs (no interaction is included yet, but will be automatically created using
the two antecedent constructs).}
}
\description{
WARNING: This is an experimental feature demonstration only - do not use it for publication purposes
}
\details{
The \code{PLSc_interact} function calculates the consistent PLS path coefficients and loadings for
a common factor model with a simple interaction. It returns a \code{seminr_model} containing the
adjusted and consistent path coefficients and loadings for a common factor model.
}
\examples{


#seminr syntax for creating measurement model
mobi_mm <- constructs(
             reflective("Image",        multi_items("IMAG", 1:5)),
             reflective("Expectation",  multi_items("CUEX", 1:3)),
             reflective("Satisfaction", multi_items("CUSA", 1:3))
           )
#seminr syntax for creating structural model
mobi_sm <- relationships(
  paths(from = "Image",        to = "Satisfaction"),
  paths(from = "Expectation",  to = "Satisfaction")
)

mobi_pls <- estimate_pls(data = mobi,
                         measurement_model = mobi_mm,
                         structural_model = mobi_sm)

interaction_mobi_plsC <- PLSc_interact(mobi_pls)

#path coefficients, R^2 and loadings
mobi_plsc_interact$path_coef
mobi_plsc_interact$rSquared
mobi_plsc_interact$outer_loadings

}
\seealso{
\code{\link{relationships}} \code{\link{constructs}} \code{\link{paths}} \code{\link{interactions}}
         \code{\link{bootstrap_model}}
}

