% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plspredict.R
\name{predict_DA}
\alias{predict_DA}
\title{Predictive Scheme}
\usage{
predict_DA(smMatrix, path_coef, construct_scores)
}
\arguments{
\item{smMatrix}{is the \code{structural_model} - a source-to-target matrix representing the inner/structural model,
generated by \code{relationships} generated by SEMinR.}

\item{path_coef}{is the Path Coefficients matrix from a SEMinR model.}

\item{construct_scores}{is the matrix of construct scores generated by SEMinR.}
}
\description{
\code{predict_EA} and \code{predict_DA} specify the predictive scheme to be used in the generation of the
predictions. EA refers to Earliest Antecedents nad DA to Direct Antecedents.
}
