% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_summary.R
\name{print.seminr_pls_mga}
\alias{print.seminr_pls_mga}
\title{Summary function for PLS-MGA}
\usage{
\method{print}{seminr_pls_mga}(x, na.print = ".", digits = 3, ...)
}
\description{
Summary function for PLS-MGA
}
