% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{print_paths}
\alias{plot_scores}
\alias{print_paths}
\title{Functions for reporting the Path Coefficients and R2 of endogenous constructs
and for generating a scatterplot matrix of factor scores.}
\usage{
print_paths(model, na.print=".", digits=2)

plot_scores(fitted_model, factors=NULL)
}
\arguments{
\item{seminr_model}{The PLS model estimated by simplePLS \code{seminr}. The estimated model
returned by the \code{estimate_pls} or \code{bootstrap_model} methods.}

\item{na.print}{A \code{character} substituting values not to be printed.
If not specified, default = "."}

\item{digits}{A \code{numeric} minimum number of significant digits. If not
specified, default is "2".}

\item{factors}{a \code{list} indicating which factors to report. If not
specified, all factors are graphed and returned.}
}
\description{
\code{print_paths} generates an easy to read table reporting path coefficients
  and R2 values for endogenous constructs.\code{plot_scores} generates a
  scatterplot matrix of each factor's scores against every other factor's scores.
}
\details{
These functions generate an easy to read table reporting path coefficients
  and R2 values for endogenous constructs or a scatterplot matrix of factor
  scores.
}
\examples{
data("mobi", package = "semPLS")
mobi_pls <- estimate_pls(data = mobi,
                           measurement_model = mobi_mm,
                           structural_model = mobi_sm)
print_paths(mobi_pls)
plot_scores(mobi_pls)

# Estimate model with bootstrapped significance
mobi_pls <- bootstrap_model(data = mobi,
                           measurement_model = mobi_mm,
                           structural_model = mobi_sm,
                           nboot = 200)

print_paths(mobi_pls)
plot_scores(mobi_pls)

}

