% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{psignr}
\alias{psignr}
\title{Format p values for the output and removes trailing numbers when p > .10}
\usage{
psignr(pvals, sig.limit = c(0.05, 0.01, 0.001), html = FALSE)
}
\arguments{
\item{pvals}{A vector with p-values}

\item{sig.limit}{The thresholds for adding an additional asterisk}

\item{html}{whether to use html coded output}
}
\value{
A string of asterisks

pvalr(c(0.432, 0.05, 0.00001))
}
\description{
Format p values for the output and removes trailing numbers when p > .10
}
\keyword{internal}
