% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{pvalr}
\alias{pvalr}
\title{Format p values for the output and removes trailing numbers when p > .10}
\usage{
pvalr(pvals, sig.limit = 0.001, digits = 3, html = FALSE)
}
\arguments{
\item{pvals}{A vector with p-values}

\item{sig.limit}{The lowest threshold for full reporting}

\item{digits}{the amount of digits to report when sig.limit < p < .10}

\item{html}{whether to use html coded output}
}
\value{
A string formatted p-value including equal and less-than sign

pvalr(c(0.432, 0.05, 0.00001))
}
\description{
Format p values for the output and removes trailing numbers when p > .10
}
\keyword{internal}
