% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{report_missing}
\alias{report_missing}
\title{Function to report how missing data was handled and how much was missing.}
\usage{
report_missing(seminr_model)
}
\arguments{
\item{seminr_model}{An estimated seminr PLS model.}
}
\description{
Function to report how missing data was handled and how much was missing.
}
