% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pls.R
\name{rerun}
\alias{rerun}
\title{Reruns a previously specified seminr model/analysis}
\usage{
rerun(x, ...)
}
\arguments{
\item{x}{An estimated seminr_model object - refer to specific rerun methods}

\item{...}{Any parameters to change during the rerun.}
}
\value{
A re-estimated model of the same class
}
\description{
Reruns a previously specified seminr model/analysis
}
\seealso{
\code{\link{rerun.pls_model}}
}
