% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_reliability.R
\name{rhoC_AVE.cbsem_model}
\alias{rhoC_AVE.cbsem_model}
\title{seminr rhoC_AVE() function}
\usage{
\method{rhoC_AVE}{cbsem_model}(x, constructs = NULL)
}
\arguments{
\item{x}{Estimated \code{cbsem_model} object.}

\item{constructs}{Vector containing the names of the constructs to calculate rhoC and AVE for; if NULL, all constructs are used.}
}
\value{
A matrix containing the rhoC and AVE metrics for each construct.
}
\description{
Get rhoC and AVE for a CBSEM model estimated with \code{estimate_cbsem}
}
