% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_reliability.R
\name{rhoC_AVE}
\alias{rhoC_AVE}
\title{seminr rhoC_AVE() function}
\usage{
rhoC_AVE(x, constructs = NULL)
}
\arguments{
\item{x}{Estimated \code{seminr_model} object.}

\item{constructs}{Vector containing the names of the constructs to calculate rhoC and AVE for; if NULL, all constructs are used.}
}
\value{
A matrix containing the rhoC and AVE metrics for each construct.
}
\description{
Get rhoC and AVE for a CFA model estimated with \code{estimate_pls}, \code{estimate_cbsem} or \code{estimate_cfa}.
Dillon-Goldstein's Rho as per: Dillon, W. R, and M. Goldstein. 1987. Multivariate Analysis: Methods
and Applications. Biometrical Journal 29 (6).
Average Variance Extracted as per:  Fornell, C. and D. F. Larcker (February 1981). Evaluating
structural equation models with unobservable variables and measurement error, Journal of Marketing Research, 18, pp. 39-5
}
