% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seminr_plot.R
\name{seminr_plot}
\alias{seminr_plot}
\title{Generates dot-Code from a model. With the help of the DiagrammR package this
code can then be plotted in various contexts.}
\usage{
seminr_plot(model, title = "", theme = NULL)
}
\arguments{
\item{model}{The PLS-model used for plotting.}

\item{title}{an optional title for the plot}

\item{theme}{an optional theme object}
}
\value{
dot-Code
}
\description{
Current limitations:
- Only plots PLS Models
- no higher order constructs
- No interaction terms
}
\examples{
mobi <- mobi

#seminr syntax for creating measurement model
mobi_mm <- constructs(
             reflective("Image",        multi_items("IMAG", 1:5)),
             reflective("Expectation",  multi_items("CUEX", 1:3)),
             reflective("Quality",      multi_items("PERQ", 1:7)),
             reflective("Value",        multi_items("PERV", 1:2)),
             reflective("Satisfaction", multi_items("CUSA", 1:3)),
             reflective("Complaints",   single_item("CUSCO")),
             reflective("Loyalty",      multi_items("CUSL", 1:3))
           )
#seminr syntax for creating structural model
mobi_sm <- relationships(
  paths(from = "Image",        to = c("Expectation", "Satisfaction", "Loyalty")),
  paths(from = "Expectation",  to = c("Quality", "Value", "Satisfaction")),
  paths(from = "Quality",      to = c("Value", "Satisfaction")),
  paths(from = "Value",        to = c("Satisfaction")),
  paths(from = "Satisfaction", to = c("Complaints", "Loyalty")),
  paths(from = "Complaints",   to = "Loyalty")
)

mobi_pls <- estimate_pls(data = mobi,
                         measurement_model = mobi_mm,
                         structural_model = mobi_sm)

# generate dot-Notation
res <- seminr_plot(mobi_pls)

\dontrun{
DiagrammR::grViz(res)
}
}
