% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_defaults.R
\name{seminr_theme_academic}
\alias{seminr_theme_academic}
\title{A theme function for a basic b/w theme}
\usage{
seminr_theme_academic(
  plot.title.fontsize = 24,
  mm.node.label.fontsize = 8,
  sm.node.label.fontsize = 12,
  mm.edge.label.fontsize = 7,
  sm.edge.label.fontsize = 9
)
}
\arguments{
\item{plot.title.fontsize}{Title font size}

\item{mm.node.label.fontsize}{Font size for measurement variables}

\item{sm.node.label.fontsize}{Font size for constructs}

\item{mm.edge.label.fontsize}{Font size for measurement model edges}

\item{sm.edge.label.fontsize}{Font size for path edges}
}
\value{
a theme object
}
\description{
A theme function for a basic b/w theme
}
