% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_defaults.R
\name{seminr_theme_smart}
\alias{seminr_theme_smart}
\alias{seminr_theme_default}
\title{A colored theme}
\usage{
seminr_theme_smart(
  plot.title.fontsize = 24,
  mm.node.label.fontsize = 8,
  sm.node.label.fontsize = 12,
  mm.edge.label.fontsize = 7,
  sm.edge.label.fontsize = 9
)

seminr_theme_default(
  plot.title.fontsize = 24,
  mm.node.label.fontsize = 8,
  sm.node.label.fontsize = 12,
  mm.edge.label.fontsize = 7,
  sm.edge.label.fontsize = 9,
  construct.reflective.shape = "ellipse",
  construct.compositeA.shape = "hexagon",
  construct.compositeB.shape = "hexagon",
  construct.reflective.arrow = "backward",
  construct.compositeA.arrow = "backward",
  construct.compositeB.arrow = "forward",
  ...
)
}
\arguments{
\item{plot.title.fontsize}{Title font size}

\item{mm.node.label.fontsize}{Font size for measurement variables}

\item{sm.node.label.fontsize}{Font size for constructs}

\item{mm.edge.label.fontsize}{Font size for measurement model edges}

\item{sm.edge.label.fontsize}{Font size for path edges}

\item{construct.reflective.shape}{Shape of reflective constructs}

\item{construct.compositeA.shape}{Shape of composite constructs mode A}

\item{construct.compositeB.shape}{Shape of composite constructs mode B}

\item{construct.reflective.arrow}{Direction of arrows of reflective constructs}

\item{construct.compositeA.arrow}{Direction of arrows of composite constructs mode A}

\item{construct.compositeB.arrow}{Direction of arrows of composite constructs mode B}

\item{...}{Other parameters for the seminr_theme_create function}
}
\value{
a theme object
}
\description{
A colored theme

A theme function for a modern approach of visualizing PLS models in b/w
}
