% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_defaults.R
\name{seminr_theme_smartpls}
\alias{seminr_theme_smartpls}
\title{A SmartPLS inspired theme}
\usage{
seminr_theme_smartpls(
  title_fontsize = 24,
  construct_fontsize = 14,
  item_fontsize = 8,
  path_fontsize = 12,
  mm_fontsize = 8
)
}
\arguments{
\item{title_fontsize}{Title font size}

\item{construct_fontsize}{Font size for constructs}

\item{item_fontsize}{Font size for measurement variables}

\item{path_fontsize}{Font size for path edges}

\item{mm_fontsize}{Font size for measurement model}
}
\value{
the theme object
}
\description{
A SmartPLS inspired theme
}
