% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lastplot.R
\name{set_last_seminr_plot}
\alias{set_last_seminr_plot}
\title{Set the last plot to be fetched by lastplot()}
\usage{
set_last_seminr_plot(value)
}
\description{
Set the last plot to be fetched by lastplot()
}
\seealso{
[last_plot()]
}
\keyword{internal}
