% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_constructs.R
\name{single_item}
\alias{single_item}
\title{Single-item measurement model specification}
\usage{
single_item(item)
}
\arguments{
\item{item}{Name of item}
}
\description{
\code{single_item} specifies a single item name to be assigned to a construct.
}
\examples{
  mobi_mm <- constructs(
    composite("Image",        multi_items("IMAG", 1:5), weights = correlation_weights),
    composite("Expectation",  multi_items("CUEX", 1:3), weights = mode_A),
    composite("Quality",      multi_items("PERQ", 1:7), weights = regression_weights),
    composite("Value",        single_item("PERV1"))
  )
}
\seealso{
See \code{\link{multi_items}}
}
