% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_models.R
\name{specify_model}
\alias{specify_model}
\title{seminr specify_model() function}
\usage{
specify_model(
  measurement_model,
  structural_model = NULL,
  item_associations = NULL
)
}
\arguments{
\item{measurement_model}{An optional \code{measurement_model} object representing the outer/measurement model,
as generated by \code{constructs}.}

\item{structural_model}{An optional \code{smMatrix} object representing the inner/structural model,
as generated by \code{relationships}.}

\item{item_associations}{An item-to-item matrix representing error
covariances that are freed for estimation.
This matrix is created by \code{associations()}, or defaults to NULL
(no inter-item associations).}
}
\description{
Combines model components together into a single \code{specified_model} object for estimation functions
}
\seealso{
\code{\link{estimate_pls}} \code{\link{estimate_cbsem}} \code{\link{estimate_cfa}}
}
