% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{SRMR}
\alias{SRMR}
\title{Function to calculate the SRMR of a model}
\usage{
SRMR(seminr_model)
}
\arguments{
\item{seminr_model}{is the PLS model estimated by SEMinR to calculate SRMR for.}
}
\description{
\code{SRMR} calculate the saturated and estimated SRMR of a estimated SEMinR model.
}
\references{
Henseler, J., Hubona, G., & Ray, P. A. (2016). Using PLS path modeling in new technology research: updated guidelines. Industrial management & data systems, 116(1), 2-20.
}
