% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_safe.R
\name{standardize_safely}
\alias{standardize_safely}
\title{Standardize (scale) a matrix/df and report interpretable errors}
\usage{
standardize_safely(x)
}
\arguments{
\item{x}{vector, data.frame, or matrix}
}
\value{
scaled object as returned by \code{scale} function
}
\description{
Standardize (scale) a matrix/df and report interpretable errors
}
