% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{structure}
\alias{paths}
\alias{structure}
\title{Structural specification functions for seminr package}
\usage{
structure(paths(from = "source factor", to = "destination factor"))
}
\arguments{
\item{from}{The source factor of a structural path}

\item{to}{The destination factor of a structural path}

\item{paths}{The function \code{paths} that specifies the source and destination factors
for each of the model's structural paths.}
}
\description{
\code{paths} creates the structural paths of a PLS SEM model and \code{structure} generates
the matrix of paths for use by PLS modelling packages such as semPLS and simplePLS.
}
\examples{
mobi_sm <- structure(
             paths(from = "Image",        to = c("Expectation", "Satisfaction", "Loyalty")),
             paths(from = "Expectation",  to = c("Quality", "Value", "Satisfaction")),
             paths(from = "Quality",      to = c("Value", "Satisfaction")),
             paths(from = "Value",        to = c("Satisfaction")),
             paths(from = "Satisfaction", to = c("Complaints", "Loyalty")),
             paths(from = "Complaints",   to = "Loyalty")
           )

}

