% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_lavaan.R
\name{summarize_cb_measurement}
\alias{summarize_cb_measurement}
\title{Creates summary statistics for a cbsem object for summary and print functions}
\usage{
summarize_cb_measurement(object)
}
\description{
Creates summary statistics for a cbsem object for summary and print functions
}
