% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bootstrap_full.R
\name{summary.boot_seminr_model_full}
\alias{summary.boot_seminr_model_full}
\title{Summary function for the bootstrap full model}
\usage{
\method{summary}{boot_seminr_model_full}(boot_object, alpha = 0.05, ...)
}
\description{
Summary function for the bootstrap full model
}
