% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{test_path_diff}
\alias{test_path_diff}
\title{Tests the difference between two path coefficients in a bootstrapped model
NOTE: Experimental features}
\usage{
test_path_diff(path(from="SCVEND", to="TTSE"),
                 path(from="SCVEND", to="INER"),
                 boot_pls)
}
\arguments{
\item{path1}{A \code{vector} containing \code{from} and \code{to} strings of construct names defining the start and end of that path.}

\item{boot_pls}{Bootstrapped SEMinR model (\code{boot_seminr_model} object)}
}
\description{
Tests the difference between two path coefficients in a bootstrapped model
NOTE: Experimental features
}
