% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library-traverse.R
\name{traverse_names}
\alias{traverse_names}
\title{Traverses tree structures of nested lists to create
string representation of structure. Useful to compare list structures
in test cases. (see cbsem summary tests for example)}
\usage{
traverse_names(list_tree, level = 0)
}
\description{
Traverses tree structures of nested lists to create
string representation of structure. Useful to compare list structures
in test cases. (see cbsem summary tests for example)
}
