% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library-errors-warnings.R
\name{try_or_stop}
\alias{try_or_stop}
\title{expr: expression to try
context_msg: context of warning or error (e.g., "run CFA in Lavaan")
             use a verb to start the message (e.g., "run...", "execute...")}
\usage{
try_or_stop(expr, context_msg = NULL)
}
\description{
expr: expression to try
context_msg: context of warning or error (e.g., "run CFA in Lavaan")
             use a verb to start the message (e.g., "run...", "execute...")
}
