% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstandardize.R
\name{unstandardize_scores}
\alias{unstandardize_scores}
\title{Rescales standardized construct scores back into same units of items.
Inspired by \code{rescale()} function of \code{plspm} package}
\usage{
unstandardize_scores(pls_model)
}
\description{
Rescales standardized construct scores back into same units of items.
Inspired by \code{rescale()} function of \code{plspm} package
}
