% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{use_construct_weights}
\alias{use_construct_weights}
\title{Should a construct type use weights or loadings}
\usage{
use_construct_weights(theme, construct_type)
}
\arguments{
\item{theme}{The theme to use}

\item{construct_type}{the construct type to test}
}
\value{
TRUE if weights should be used, FALSE if loadings
}
\description{
Should a construct type use weights or loadings
}
\keyword{internal}
