% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_mmMatrix.R
\name{as.reflective}
\alias{as.reflective}
\title{Converts all contructs of a measurement model, or just a single construct
 into reflective factors.}
\usage{
as.reflective(x, ...)
}
\arguments{
\item{x}{A measurement model defined by \code{\link{constructs}} 
or a single composite construct defined by \code{\link{composite}}}
}
\description{
Converts all contructs of a measurement model, or just a single construct
 into reflective factors.
}
\examples{

mobi_mm <- constructs(
  composite("Image",        multi_items("IMAG", 1:5)),
  composite("Expectation",  multi_items("CUEX", 1:3)),
  composite("Value",        multi_items("PERV", 1:2))
)

fa <- estimate_cfa(as.reflective(mobi_mm))

}
\seealso{
\code{\link{as.reflective.measurement_model}},
  \code{\link{as.reflective.construct}}
}
