% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{browse_plot}
\alias{browse_plot}
\title{Open Edotor graphViz Website with the preloaded in the Browser}
\usage{
browse_plot(model)
}
\arguments{
\item{model}{A SEMinR Model}
}
\description{
Open Edotor graphViz Website with the preloaded in the Browser
}
\examples{
\dontrun{
browse_plot(model)
}
}
