% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_mmMatrix.R
\name{construct_items}
\alias{construct_items}
\title{Get names of all items of a construct in a measurement model}
\usage{
construct_items(construct)
}
\description{
Get names of all items of a construct in a measurement model
}
