% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_smMatrix.R
\name{construct_names}
\alias{construct_names}
\title{Get all unique construct names in a structural model}
\usage{
construct_names(smMatrix)
}
\description{
Get all unique construct names in a structural model
}
