% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{cor_rsq}
\alias{cor_rsq}
\title{Returns R-sq of a dv given correlation matrix of ivs, dv}
\usage{
cor_rsq(cor_matrix, dv_name, iv_names)
}
\arguments{
\item{cor_matrix}{A correlation matrix that includes ivs and dv}

\item{dv_name}{Character string of dependent variable}

\item{iv_names}{Vector of character strings for independent variables}
}
\description{
Returns R-sq of a dv given correlation matrix of ivs, dv
}
\examples{
cors <- cbsem_summary$descriptives$correlations$constructs
cor_rsq(cors, dv_name = "Value", iv_names = c("Image", "Quality"))

}
