% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{cor_vifs}
\alias{cor_vifs}
\title{Computes Variance Inflation Factor (VIF) for named independent variables
in a correlation table}
\usage{
cor_vifs(cor_matrix, iv_names)
}
\description{
Computes Variance Inflation Factor (VIF) for named independent variables
in a correlation table
}
