% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{dot_graph}
\alias{dot_graph}
\title{Convert a seminr model to Graphviz representation}
\usage{
dot_graph(model, theme = create_theme())
}
\arguments{
\item{model}{Model created with \code{seminr}.}

\item{theme}{Theme created with \code{\link{create_theme}}.}
}
\value{
The path model as a formatted string in dot language.
}
\description{
Convert a seminr model to Graphviz representation
}
