% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_dot.R
\name{dot_graph.structural_model}
\alias{dot_graph.structural_model}
\title{Convert a seminr measurement model to a Graphviz representation}
\usage{
\method{dot_graph}{structural_model}(model, title = "", theme = NULL, ...)
}
\arguments{
\item{model}{Model created with \code{seminr}.}

\item{title}{An optional title for the plot}

\item{theme}{Theme created with \code{\link{seminr_theme_create}}.}

\item{...}{Unused}
}
\description{
With the help of the \code{DiagrammeR} package this code can then be plotted in
various contexts.
}
\examples{
# - - - - - - - - - - - - - - - -
# Example for plotting a measurement model
mobi_mm <- constructs(
             reflective("Image",        multi_items("IMAG", 1:5)),
             reflective("Expectation",  multi_items("CUEX", 1:3)),
             reflective("Quality",      multi_items("PERQ", 1:7)),
             reflective("Value",        multi_items("PERV", 1:2)),
             reflective("Satisfaction", multi_items("CUSA", 1:3)),
             reflective("Complaints",   single_item("CUSCO")),
             reflective("Loyalty",      multi_items("CUSL", 1:3))
           )
res <- dot_graph(mobi_mm, title = "Preview measurement model")

\dontrun{
DiagrammeR::grViz(res)
}
}
