% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cbsem.R
\name{estimate_cbsem}
\alias{estimate_cbsem}
\title{seminr estimate_cbsem() function}
\usage{
estimate_cbsem(data, measurement_model, structural_model)
}
\arguments{
\item{data}{A \code{dataframe} containing the indicator measurement data.}

\item{measurement_model}{A source-to-target matrix representing the outer/measurement model,
generated by \code{constructs}. Note that only reflective constructs are supported.}

\item{structural_model}{A source-to-target matrix representing the inner/structural model,
generated by \code{relationships}.}
}
\description{
The \code{seminr} package provides a natural syntax for researchers to describe
structural equation models.
}
\examples{
mobi <- mobi

#seminr syntax for creating measurement model
mobi_mm <- constructs(
             reflective("Image",        multi_items("IMAG", 1:5)),
             reflective("Expectation",  multi_items("CUEX", 1:3)),
             reflective("Quality",      multi_items("PERQ", 1:7)),
             reflective("Value",        multi_items("PERV", 1:2)),
             reflective("Satisfaction", multi_items("CUSA", 1:3)),
             reflective("Complaints",   single_item("CUSCO")),
             reflective("Loyalty",      multi_items("CUSL", 1:3))
           )

#seminr syntax for freeing up item-item covariances
mobi_am <- associations(
             item_errors(c("PERQ1", "PERQ2"), "CUEX3"),
             item_errors("IMAG1", "CUEX2")
           )

#seminr syntax for creating structural model
mobi_sm <- relationships(
  paths(from = "Image",        to = c("Expectation", "Satisfaction", "Loyalty")),
  paths(from = "Expectation",  to = c("Quality", "Value", "Satisfaction")),
  paths(from = "Quality",      to = c("Value", "Satisfaction")),
  paths(from = "Value",        to = c("Satisfaction")),
  paths(from = "Satisfaction", to = c("Complaints", "Loyalty")),
  paths(from = "Complaints",   to = "Loyalty")
)

# Estimate model and get results
cbsem <- estimate_cbsem(mobi, mobi_mm, mobi_sm, mobi_am)
cbsem$results
}
\seealso{
\code{\link{as.reflective}} \code{\link{all.reflective}}
         \code{\link{relationships}} \code{\link{constructs}}
         \code{\link{paths}} \code{\link{interaction_term}}
         \code{\link{bootstrap_model}}
}
