% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cbsem.R
\name{estimate_cfa}
\alias{estimate_cfa}
\title{#' @examples
mobi <- mobi}
\usage{
estimate_cfa(
  data,
  measurement_model,
  item_associations = NULL,
  estimator = "MLR",
  ...
)
}
\description{
#seminr syntax for creating measurement model
mobi_mm <- constructs(
reflective("Image",          multi_items("IMAG", 1:5)),
  reflective("Expectation",  multi_items("CUEX", 1:3)),
  reflective("Quality",      multi_items("PERQ", 1:7))
)
}
\details{
#seminr syntax for freeing up item-item covariances
mobi_am <- associations(
             item_errors(c("PERQ1", "PERQ2"), "CUEX3"),
             item_errors("IMAG1", "CUEX2")
           )

mobi_cfa <- estimate_cfa(mobi, mobi_mm, mobi_am)
}
