% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{extract_sm_nodes}
\alias{extract_sm_nodes}
\title{Helper function that applies formatting to each construct}
\usage{
extract_sm_nodes(model, theme, structure_only = FALSE)
}
\arguments{
\item{model}{the model to use}

\item{theme}{the theme to use}

\item{structure_only}{is this called in a structure_only model}
}
\value{
Returns a string of the structural model in dot notation.
}
\description{
Helper function that applies formatting to each construct
}
