% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_validity.R
\name{fl_criteria_table}
\alias{fl_criteria_table}
\title{fl_criteria_table can be used to generate simple and effective table for checking Fornell Larcker criteria.
Fornell, C., & Larcker, D. F. (1981). Evaluating structural equation models with unobservable variables and measurement error. Journal of marketing research, 18(1), 39-50.}
\usage{
fl_criteria_table(seminr_model)
}
\description{
fl_criteria_table can be used to generate simple and effective table for checking Fornell Larcker criteria.
Fornell, C., & Larcker, D. F. (1981). Evaluating structural equation models with unobservable variables and measurement error. Journal of marketing research, 18(1), 39-50.
}
