% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_dot.R
\name{get_global_style}
\alias{get_global_style}
\title{Get dot string for global theme options}
\usage{
get_global_style(theme)
}
\arguments{
\item{theme}{a theme}
}
\description{
Get dot string for global theme options
}
\keyword{internal}
