% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{get_mm_edge_style}
\alias{get_mm_edge_style}
\title{individual styles for measurement model edges}
\usage{
get_mm_edge_style(theme, forward)
}
\arguments{
\item{theme}{the theme to use}

\item{forward}{Forward direction?}
}
\description{
individual styles for measurement model edges
}
