% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{influencer_data}
\alias{influencer_data}
\title{Measurement Instrument for the Influencer Model}
\format{
A data frame with 250 rows and 24 variables:
\describe{

  \item{SIC1}{The influencer reflects who I am.}
  \item{SIC2}{I can identify with the influencer.}
  \item{SIC3}{I feel a personal connection to the influencer.}
  \item{SIC4}{I (can) use the influencer to communicate who I am to other people.}
  \item{SIC5}{I think the influencer (could) help(s) me become the type of person I want to be.}
  \item{SIC6}{I consider the influencer to be "me".}
  \item{SIC7}{The influencer suits me well.}
  \item{SIC_G}{My personality and the personality of the influencer relate accordingly to one another.}
  \item{PQ1}{The product has excellent quality.}
  \item{PQ2}{The product looks to be reliable and durable.}
  \item{PQ3}{The product will have fewer problems.}
  \item{PQ4}{The product has excellent quality features.}
  \item{PL1}{I dislike the product (reverse coded).}
  \item{PL2}{The product is appealing to me.}
  \item{PL3}{The presented product raises a positive feeling in me.}
  \item{PL4}{The product is interesting to me.}
  \item{PI1}{It is very likely that I will purchase this product.}
  \item{PI2}{I will purchase this product the next time I need it.}
  \item{PI3}{I would definitely try the product out.}
  \item{PI4}{I would recommend this product to my friends.}
  \item{PI5}{I am willing to purchase this product.}
  \item{PIC1}{The influencer is qualified.}
  \item{PIC2}{The influencer is competent.}
  \item{PIC3}{The influencer is an expert.}
  \item{PIC4}{The influencer is experienced.}
  \item{PIC5}{The influencer is knowledgeable.}
  \item{WTP}{Please state your willingness to pay (in Euro) for the presented product.}
}
}
\usage{
influencer_data
}
\description{
The data set is used as measurement instrument for the Influencer Model which is used in
Partial Least Squares Structural Equation Modeling (PLS-SEM) Using R - A Workbook (2021)
Hair, J.F. (Jr), Hult, T.M., Ringle, C.M., Sarstedt, M., Danks, N.P., and Ray, S.
}
\details{
The data frame influencer_data contains the observed data for the model specified in the Influencer Model.
}
\examples{
data("influencer_data")

}
\keyword{datasets}
