% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interaction_ortho}
\alias{interaction_ortho}
\alias{interaction_ortho,}
\alias{interaction_scaled}
\title{\code{interaction_ortho} creates interaction measurement items by multipying all combination of factor items.}
\usage{
# Create two new interactions (factor1.factor2 and factor1.factor3) using

 # orthogonalization approach as per Henseler & CHin (2010):
 interactions( interaction_ortho("factor1", "factor2"),
           interaction_ortho("factor1", "factor3"))
}
\arguments{
\item{...}{Interaction Combinations as generated by the \code{interaction_ortho} method.}
}
\description{
This function automatically generates interaction measurement items for a PLS SEM.
}
\examples{
mobi <- mobi

mobi_xm <- interactions(
  interaction_ortho("Image", "Expectation"),
  interaction_ortho("Image", "Value")
)

mobi_pls <- estimate_pls(mobi, mobi_mm, mobi_xm, mobi_sm)
print_paths(mobi_pls)

}

