% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{mean_replacement}
\alias{mean_replacement}
\title{Function to clean data of omitted values by mean replacement}
\usage{
mean_replacement(data)
}
\arguments{
\item{data}{A dataset to be used for estimating a SEMinR model}
}
\value{
A dataset with all missing values replaced with column means
}
\description{
The \code{seminr} package provides a natural syntax for researchers to describe PLS
structural equation models.
}
\details{
\code{mean_replacement} provides the verb for replacing all omitted values (NA only) in the dataset with
the mean of the variable.
}
\references{
Hair, J. F., Hult, G. T. M., Ringle, C. M., and Sarstedt, M. (2017). A Primer on Partial Least Squares
 Structural Equation Modeling (PLS-SEM), 2nd Ed., Sage: Thousand Oaks.
}
