% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement.R
\name{measure}
\alias{measure}
\title{Measurement functions}
\usage{
measurement_model <- measure(
                       reflect("factor1", multi_items("fac1", 1:3)),
                       reflect("factor2", multi_items("fac2", 1:3)),
                       )
}
\arguments{
\item{...}{Reflective or Formative factors as generated by the \code{reflect()}
or \code{form()}}
}
\description{
\code{measure} creates the factors from measurement items by assigning the
relevant items to each factor and specifying formative or reflective type
}
\details{
This function conveniently maps measurement items to factors using
root name, numbers, and affixes with explicit definition of formative
or reflective structure
}
\examples{
  mobi_mm <- measure(
   reflect("Image",        multi_items("IMAG", 1:5)),
   reflect("Expectation",  multi_items("CUEX", 1:3)),
   reflect("Quality",      multi_items("PERQ", 1:7)),
   reflect("Value",        multi_items("PERV", 1:2)),
   reflect("Satisfaction", multi_items("CUSA", 1:3)),
   reflect("Complaints",   single_item("CUSCO")),
   reflect("Loyalty",      multi_items("CUSL", 1:3))
   )
}
\seealso{
See \code{\link{form}}, \code{\link{reflect}}, \code{\link{multi_items}}
  and \code{\link{single_item}}
}

