% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot.reliability_table}
\alias{plot.reliability_table}
\title{Plot Reliability Tablecan be used to generate simple and effective bar plots for reliability stats.}
\usage{
\method{plot}{reliability_table}(object)
}
\description{
Plot Reliability Tablecan be used to generate simple and effective bar plots for reliability stats.
}
