% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Generating graphviz code for a model object.}
\usage{
plot_model(
  model,
  use_outer_weights = FALSE,
  title = "",
  title_font = "helvetica",
  title_size = 24,
  theme = create_theme()
)
}
\arguments{
\item{model}{A model object generated from seminr (can be a PLS, PLSc, lavaan or a bootstrapped model).}

\item{use_outer_weights}{determines whether to use weights or factor loadings}

\item{title}{The title of the diagram}

\item{title_font}{the font to use in the title (default: "helvetica")}

\item{title_size}{the font size of the title (default: 24pt)}

\item{theme}{a seminr visualization theme. Generated by the create_theme method}
}
\value{
A graphviz text that can be rendered using the DiagrammeR package
}
\description{
Generating graphviz code for a model object.
}
\examples{
mobi <- mobi

#seminr syntax for creating measurement model
mobi_mm <- constructs(
             reflective("Image",        multi_items("IMAG", 1:5)),
             reflective("Expectation",  multi_items("CUEX", 1:3)),
             reflective("Quality",      multi_items("PERQ", 1:7)),
             reflective("Value",        multi_items("PERV", 1:2)),
             reflective("Satisfaction", multi_items("CUSA", 1:3)),
             reflective("Complaints",   single_item("CUSCO")),
             reflective("Loyalty",      multi_items("CUSL", 1:3))
           )
#seminr syntax for creating structural model
mobi_sm <- relationships(
  paths(from = "Image",        to = c("Expectation", "Satisfaction", "Loyalty")),
  paths(from = "Expectation",  to = c("Quality", "Value", "Satisfaction")),
  paths(from = "Quality",      to = c("Value", "Satisfaction")),
  paths(from = "Value",        to = c("Satisfaction")),
  paths(from = "Satisfaction", to = c("Complaints", "Loyalty")),
  paths(from = "Complaints",   to = "Loyalty")
)

mobi_pls <- estimate_pls(data = mobi,
                         measurement_model = mobi_mm,
                         structural_model = mobi_sm)

plot_model(mobi_pls)
}
