% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistent_interaction.R
\name{PLSc_interact}
\alias{PLSc_interact}
\title{seminr Consistent Interaction Function}
\usage{
PLSc_interact(seminr.model)
}
\arguments{
\item{seminr.model}{A \code{seminr.model} containing the estimated seminr model with only the two
antecedent constructs (no interaction is included yet, but will be automatically created using
the two antecedent constructs).}
}
\description{
The \code{PLSc_interact} function calculates the consistent PLS path coefficients and loadings for
a common factor model with a simple interaction. It returns a \code{seminr.model} containing the
adjusted and consistent path coefficients and loadings for a common factor model.
}
\examples{
mobi <- mobi

#seminr syntax for creating measurement model
mobi_mm <- constructs(
             reflective("Image",        multi_items("IMAG", 1:5)),
             reflective("Expectation",  multi_items("CUEX", 1:3)),
             reflective("Satisfaction", multi_items("CUSA", 1:3))
           )
#seminr syntax for creating structural model
mobi_sm <- relationships(
  paths(from = "Image",        to = "Satisfaction"),
  paths(from = "Expectation",  to = "Satisfaction")
)

mobi.pls <- estimate_pls(data = mobi,
                         measurement_model = mobi_mm,
                         structural_model = mobi_sm)

mobi.pls.consistent.interaction <- PLSc_interact(mobi.pls)

#path coefficients, R^2 and loadings
mobi.pls.consistent.interaction$path_coef
mobi.pls.consistent.interaction$rSquared
mobi.pls.consistent.interaction$outer_loadings

}
\seealso{
\code{\link{relationships}} \code{\link{constructs}} \code{\link{paths}} \code{\link{interactions}}
         \code{\link{bootstrap_model}}
}
