% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_summary.R
\name{print.seminr_pls_mga}
\alias{print.seminr_pls_mga}
\title{Summary function for PLS-MGA}
\usage{
\method{print}{seminr_pls_mga}(x, digits = 3, ...)
}
\arguments{
\item{x}{estimated seminr_pls_mga object}

\item{digits}{number of digits to print}

\item{...}{any further parameters for printing}
}
\description{
Summary function for PLS-MGA
}
