% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seminr-package.R
\docType{package}
\name{seminr-package}
\alias{seminr}
\alias{seminr-package}
\title{seminr: Building and Estimating Structural Equation Models}
\description{
A powerful, easy to use syntax for specifying and estimating complex Structural Equation Models. Models can be estimated using Partial Least Squares Path Modeling or Covariance-Based Structural Equation Modeling or covariance based Confirmatory Factor Analysis (Ray, Danks, and Valdez 2021 \doi{10.2139/ssrn.3900621}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sem-in-r/seminr}
  \item Report bugs at \url{https://github.com/sem-in-r/seminr/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Patrick Danks \email{nicholasdanks@hotmail.com}

Authors:
\itemize{
  \item Soumya Ray \email{soumya.ray@gmail.com} [thesis advisor]
  \item André Calero Valdez \email{andrecalerovaldez@gmail.com}
}

Other contributors:
\itemize{
  \item Juan Manuel Velasquez Estrada [contributor]
  \item James Uanhoro \email{James.uanhoro@gmail.com} [contributor]
  \item Johannes Nakayama \email{johannes.nakayama@rwth-aachen.de} [contributor]
  \item Lilian Koyan \email{lilian.kojan@rwth-aachen.de} [contributor]
  \item Laura Burbach \email{laura.burbach@rwth-aachen.de} [contributor]
  \item Arturo Heynar Cano Bejar \email{arturocano997@gmail.com} [contributor]
  \item Susanne Adler \email{susanne.adler@ovgu.de} [contributor]
}

}
\keyword{internal}
