% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{seminr_theme_create}
\alias{seminr_theme_create}
\title{Create a theme for a seminr graph visualization}
\usage{
seminr_theme_create(
  plot.title.fontsize = 24,
  plot.fontname = "helvetica",
  plot.splines = TRUE,
  plot.rounding = 3,
  plot.adj = TRUE,
  plot.specialcharacters = TRUE,
  mm.node.color = "dimgrey",
  mm.node.fill = "white",
  mm.node.label.fontsize = 8,
  mm.edge.positive.color = "dimgrey",
  mm.edge.negative.color = "dimgrey",
  mm.edge.positive.style = "solid",
  mm.edge.negative.style = "dashed",
  mm.edge.label.fontsize = 7,
  mm.edge.minlen = 1,
  mm.edge.use_outer_weights = TRUE,
  mm.edge.boot.show_t_value = FALSE,
  mm.edge.boot.show_p_value = FALSE,
  mm.edge.boot.show_p_stars = TRUE,
  mm.edge.boot.show_ci = FALSE,
  mm.edge.boot.template = edge_template_minimal(),
  sm.node.color = "black",
  sm.node.fill = "white",
  sm.node.label.fontsize = 12,
  sm.node.endo.template = node_endo_template_default(),
  sm.node.exo.template = node_exo_template_default(),
  sm.edge.boot.show_t_value = FALSE,
  sm.edge.boot.show_p_value = FALSE,
  sm.edge.boot.show_p_stars = TRUE,
  sm.edge.boot.show_ci = TRUE,
  sm.edge.boot.template = edge_template_default(),
  sm.edge.positive.color = "black",
  sm.edge.negative.color = "black",
  sm.edge.positive.style = "solid",
  sm.edge.negative.style = "dashed",
  sm.edge.label.fontsize = 9,
  sm.edge.minlen = NA,
  construct.reflective.shape = "ellipse",
  construct.compositeA.shape = "ellipse",
  construct.compositeB.shape = "ellipse",
  manifest.reflective.shape = "box",
  manifest.compositeA.shape = "box",
  manifest.compositeB.shape = "box"
)
}
\arguments{
\item{plot.title.fontsize}{Font size of the title.}

\item{plot.fontname}{Font to be used throughout the plot.}

\item{plot.splines}{Whether or not to use splines as edges.}

\item{plot.rounding}{The amount of decimals to keep for rounding.}

\item{plot.adj}{Whether or not to use adjusted r^2 in constructs}

\item{plot.specialcharacters}{Whether or not to use greek UTF-8 symbols in plots}

\item{mm.node.color}{Color of the measurement model nodes.}

\item{mm.node.fill}{Fill of the measurement model nodes.}

\item{mm.node.label.fontsize}{Font size of the measurement model node labels.}

\item{mm.edge.positive.color}{Color of the measurement model edges, when values are positive.}

\item{mm.edge.negative.color}{Color of the measurement model edges, when values are negative.}

\item{mm.edge.positive.style}{Style of the measurement model edges, when values are positive.}

\item{mm.edge.negative.style}{Style of the measurement model edges, when values are negative.}

\item{mm.edge.label.fontsize}{Font size of the measurement model edge labels.}

\item{mm.edge.minlen}{Minimum length of the measurement model edges.}

\item{mm.edge.use_outer_weights}{Whether or not to use outer weights as edge labels in the measurement model.}

\item{mm.edge.boot.show_t_value}{Should boot-strapped loadings/weights show a t-value}

\item{mm.edge.boot.show_p_value}{Should boot-strapped loadings/weights show a p-value}

\item{mm.edge.boot.show_p_stars}{Should boot-strapped loadings/weights show significance stars}

\item{mm.edge.boot.show_ci}{Should boot-strapped loadings/weights show a 95 percent confidence interval}

\item{mm.edge.boot.template}{A template string for HTML formatting of edges for loadings/weights}

\item{sm.node.color}{Color of the structural model nodes.}

\item{sm.node.fill}{Fill of the structural model nodes.}

\item{sm.node.label.fontsize}{Font size of the structural model node labels.}

\item{sm.node.endo.template}{A template string for the nodes of endogenous constructs}

\item{sm.node.exo.template}{A template string for the nodes of exogenous constructs}

\item{sm.edge.boot.show_t_value}{Should boot-strapped path coefficients show a t-value}

\item{sm.edge.boot.show_p_value}{Should boot-strapped path coefficients show a p-value}

\item{sm.edge.boot.show_p_stars}{Should boot-strapped path coefficients show significance stars}

\item{sm.edge.boot.show_ci}{Should boot-strapped path coefficients show a 95 percent confidence interval}

\item{sm.edge.boot.template}{A template string for HTML formatting of edges}

\item{sm.edge.positive.color}{Color of the structural model edges, when values are positive.}

\item{sm.edge.negative.color}{Color of the structural model edges, when values are negative.}

\item{sm.edge.positive.style}{Style of the structural model edges, when values are positive.}

\item{sm.edge.negative.style}{Style of the structural model edges, when values are negative.}

\item{sm.edge.label.fontsize}{Font size of the structural model edge labels.}

\item{sm.edge.minlen}{Minimum length of the structural model edges.}

\item{construct.reflective.shape}{Dot shape of reflective constructs}

\item{construct.compositeA.shape}{Dot shape of composite constructs using correlation weights}

\item{construct.compositeB.shape}{Dot shape of composite constructs using regression weights}

\item{manifest.reflective.shape}{Dot shape of manifest variables of reflective constructs}

\item{manifest.compositeA.shape}{Dot shape of manifest variables of composite constructs using correlation weights}

\item{manifest.compositeB.shape}{Dot shape of manifest variables of composite constructs using regression weights}
}
\value{
A \code{seminr.theme} object that can be supplied to \code{\link{dot_graph}}
}
\description{
Create a theme for a seminr graph visualization
}
